apps/worker/pyproject.toml
[project]
name = "jepx-worker"
version = "0.0.0"
description = "JEPX-Storage compute worker — ingest, ML, LSM, AI agent. Runs on Modal (Tokyo)."
requires-python = ">=3.11,<3.13"
dependencies = [
    "modal>=0.64",
    "pydantic>=2.7",
    "python-dotenv>=1.0",
    # M2 — DB seed (reference data + data dictionary).
    "psycopg[binary]>=3.2.13",
    "holidays>=0.50",
    "pyyaml>=6.0",
    # M3 — Tier 1 ingest (HTTP fetch, retries, parsing, audit).
    "httpx>=0.27",
    "tenacity>=8.4",
    "pandas>=2.2",
    "sentry-sdk>=2.0",
]

[project.optional-dependencies]
# Heavier deps land at their respective milestones:
# M5 regime:    statsmodels
# M6 vlstm:     torch, lightning, numpy, pyarrow
# M7 lsm:       numba
# M9 agent:     fastapi, openai, sqlglot
dev = [
    "ruff>=0.6",
    "mypy>=1.10",
    "pytest>=8.0",
    "types-PyYAML>=6.0",
    "pandas-stubs>=2.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["common", "ingest", "stack", "regime", "vlstm", "lsm", "backtest", "agent", "seed"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]

[tool.mypy]
python_version = "3.11"
